refactore: UI 와 로또 실행 클래스 분리 - 들여쓰기 요구사항에 맞게 수정
